requirements.txt
accelerate==0.23.0
appdirs==1.4.4
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.0
cycler==0.11.0
docker-pycreds==0.4.0
filelock==3.12.3
fonttools==4.42.1
fsspec==2023.9.0
gitdb==4.0.10
GitPython==3.1.35
huggingface-hub==0.16.4
idna==3.4
importlib-resources==6.0.1
kiwisolver==1.4.5
matplotlib==3.7.2
numpy==1.25.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
opencv-python==4.8.0.76
packaging==23.1
pandas==2.1.1
pathtools==0.1.2
Pillow==10.0.0
protobuf==4.24.3
psutil==5.9.5
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
safetensors==0.3.3
SciencePlots==2.1.0
scipy==1.11.2
sentry-sdk==1.30.0
setproctitle==1.3.2
six==1.16.0
smmap==5.0.0
tokenizers==0.13.3
torch==1.13.1+cu117
torchaudio==0.13.1
torchvision==0.14.1+cu117
tqdm==4.66.1
transformers==4.33.1
typing_extensions==4.7.1
tzdata==2023.3
urllib3==2.0.4
wandb==0.15.10
zipp==3.16.2
